Reorganise the code as per CAPI reference
