+++ cmd/capibmadm/cmd/vpc/key/list.go
@@ -29,7 +29,7 @@ import (
 	"sigs.k8s.io/cluster-api-provider-ibmcloud/cmd/capibmadm/options"
 	"sigs.k8s.io/cluster-api-provider-ibmcloud/cmd/capibmadm/pointer"
 	"sigs.k8s.io/cluster-api-provider-ibmcloud/cmd/capibmadm/printer"
-	"sigs.k8s.io/cluster-api-provider-ibmcloud/pkg/cloud/pagingutils"
+	"sigs.k8s.io/cluster-api-provider-ibmcloud/pkg/util/paging"
 )
 
 // ListCommand vpc key list command.
@@ -79,7 +79,7 @@ func listKeys(ctx context.Context) error {
 		return true, "", nil
 	}
 
-	if err = pagingutils.PagingHelper(f); err != nil {
+	if err = paging.Helper(f); err != nil {
 		return err
 	}
