+++ cmd/main.go
@@ -51,12 +51,13 @@ import (
 	powervsinfrav1 "sigs.k8s.io/cluster-api-provider-ibmcloud/api/powervs/v1beta3"
 	vpcinfrav1beta1 "sigs.k8s.io/cluster-api-provider-ibmcloud/api/vpc/v1beta1"
 	vpcinfrav1 "sigs.k8s.io/cluster-api-provider-ibmcloud/api/vpc/v1beta2"
-	"sigs.k8s.io/cluster-api-provider-ibmcloud/controllers"
-	"sigs.k8s.io/cluster-api-provider-ibmcloud/internal/webhooks/powervs"
-	"sigs.k8s.io/cluster-api-provider-ibmcloud/internal/webhooks/vpc"
+	powervscontroller "sigs.k8s.io/cluster-api-provider-ibmcloud/internal/controllers/powervs"
+	vpccontroller "sigs.k8s.io/cluster-api-provider-ibmcloud/internal/controllers/vpc"
 	"sigs.k8s.io/cluster-api-provider-ibmcloud/pkg/cloud/endpoints"
 	"sigs.k8s.io/cluster-api-provider-ibmcloud/pkg/cloud/options"
 	"sigs.k8s.io/cluster-api-provider-ibmcloud/pkg/util/record"
+	powervs "sigs.k8s.io/cluster-api-provider-ibmcloud/webhooks/powervs/admission"
+	vpc "sigs.k8s.io/cluster-api-provider-ibmcloud/webhooks/vpc/admission"
 
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	_ "k8s.io/component-base/logs/json/register"
@@ -279,9 +280,9 @@ func main() {
 	// Setup the context that's going to be used in controllers and for the manager.
 	ctx := ctrl.SetupSignalHandler()
 
-	setupReconcilers(ctx, mgr, serviceEndpoint)
 	setupWebhooks(mgr)
 	setupChecks(mgr)
+	setupReconcilers(ctx, mgr, serviceEndpoint, watchFilterValue, skipCRDMigrationPhases)
 
 	// +kubebuilder:scaffold:builder
 	setupLog.Info("starting manager")
@@ -291,8 +292,19 @@ func main() {
 	}
 }
 
-func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []endpoints.ServiceEndpoint) {
-	// Note: The kubebuilder RBAC markers above has to be kept in sync
+func setupChecks(mgr ctrl.Manager) {
+	if err := mgr.AddReadyzCheck("webhook", mgr.GetWebhookServer().StartedChecker()); err != nil {
+		setupLog.Error(err, "unable to create ready check")
+		os.Exit(1)
+	}
+	if err := mgr.AddHealthzCheck("webhook", mgr.GetWebhookServer().StartedChecker()); err != nil {
+		setupLog.Error(err, "unable to create health check")
+		os.Exit(1)
+	}
+}
+
+func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []endpoints.ServiceEndpoint, watchFilterValue string, skipCRDMigrationPhases []string) {
+	// Note: The kubebuilder RBAC markers above have to be kept in sync
 	// with the CRDs that should be migrated by this provider.
 	crdMigratorConfig := map[client.Object]crdmigrator.ByObjectConfig{
 		&powervsinfrav1.IBMPowerVSCluster{}:         {UseCache: true, UseStatusForStorageVersionMigration: true},
@@ -302,7 +314,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		&powervsinfrav1.IBMPowerVSImage{}:           {UseCache: true, UseStatusForStorageVersionMigration: true},
 	}
 
-	crdMigratorSkipPhases := make([]crdmigrator.Phase, 0, 1)
+	crdMigratorSkipPhases := make([]crdmigrator.Phase, 0, len(skipCRDMigrationPhases))
 	for _, p := range skipCRDMigrationPhases {
 		crdMigratorSkipPhases = append(crdMigratorSkipPhases, crdmigrator.Phase(p))
 	}
@@ -318,7 +330,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMVPCClusterReconciler{
+	if err := (&vpccontroller.IBMVPCClusterReconciler{
 		Client:          mgr.GetClient(),
 		Log:             ctrl.Log.WithName("controllers").WithName("IBMVPCCluster"),
 		Recorder:        mgr.GetEventRecorderFor("ibmvpccluster-controller"),
@@ -329,7 +341,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMVPCMachineReconciler{
+	if err := (&vpccontroller.IBMVPCMachineReconciler{
 		Client:          mgr.GetClient(),
 		Log:             ctrl.Log.WithName("controllers").WithName("IBMVPCMachine"),
 		Recorder:        mgr.GetEventRecorderFor("ibmvpcmachine-controller"),
@@ -340,7 +352,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMVPCMachineTemplateReconciler{
+	if err := (&vpccontroller.IBMVPCMachineTemplateReconciler{
 		Client:          mgr.GetClient(),
 		Scheme:          mgr.GetScheme(),
 		ServiceEndpoint: serviceEndpoint,
@@ -349,7 +361,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMPowerVSClusterReconciler{
+	if err := (&powervscontroller.IBMPowerVSClusterReconciler{
 		Client:           mgr.GetClient(),
 		Recorder:         mgr.GetEventRecorderFor("ibmpowervscluster-controller"),
 		ServiceEndpoint:  serviceEndpoint,
@@ -360,7 +372,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMPowerVSMachineReconciler{
+	if err := (&powervscontroller.IBMPowerVSMachineReconciler{
 		Client:           mgr.GetClient(),
 		Recorder:         mgr.GetEventRecorderFor("ibmpowervsmachine-controller"),
 		ServiceEndpoint:  serviceEndpoint,
@@ -371,7 +383,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMPowerVSMachineTemplateReconciler{
+	if err := (&powervscontroller.IBMPowerVSMachineTemplateReconciler{
 		Client: mgr.GetClient(),
 		Scheme: mgr.GetScheme(),
 	}).SetupWithManager(ctx, mgr); err != nil {
@@ -379,7 +391,7 @@ func setupReconcilers(ctx context.Context, mgr ctrl.Manager, serviceEndpoint []e
 		os.Exit(1)
 	}
 
-	if err := (&controllers.IBMPowerVSImageReconciler{
+	if err := (&powervscontroller.IBMPowerVSImageReconciler{
 		Client:          mgr.GetClient(),
 		Recorder:        mgr.GetEventRecorderFor("ibmpowervsimage-controller"),
 		ServiceEndpoint: serviceEndpoint,
@@ -421,17 +433,6 @@ func setupWebhooks(mgr ctrl.Manager) {
 	}
 	if err := (&powervs.IBMPowerVSClusterTemplate{}).SetupWebhookWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create webhook", "webhook", "IBMPowerVSClusterTemplate")
-		os.Exit(1)
-	}
-}
-
-func setupChecks(mgr ctrl.Manager) {
-	if err := mgr.AddReadyzCheck("webhook", mgr.GetWebhookServer().StartedChecker()); err != nil {
-		setupLog.Error(err, "unable to create ready check")
-		os.Exit(1)
-	}
-	if err := mgr.AddHealthzCheck("webhook", mgr.GetWebhookServer().StartedChecker()); err != nil {
-		setupLog.Error(err, "unable to create health check")
 		os.Exit(1)
 	}
 }
