+++ pkg/cloud/util/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2025 The Kubernetes Authors.
+Copyright 2026 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,8 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-/*
-Package pagingutils provides utilities for handling pagination while listing resources.
-It includes helper functions to parse pagination tokens and iterate over paginated resources.
-*/
-package pagingutils
+// Package util provides IBM Cloud utility functions shared across cloud providers.
+package util
