+++ pkg/util/paging/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Kubernetes Authors.
+Copyright 2025 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,5 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package genutil implements general utility code.
-package genutil
+// Package paging provides utilities for handling pagination while listing resources.
+// It includes helper functions to parse pagination tokens and iterate over paginated resources.
+package paging
